Fix mac build.  No standard headers in Audacity.h.
