+++ src/ShuttleGui.h
@@ -16,6 +16,7 @@
 
 #include "Audacity.h"
 
+#include "MemoryX.h"
 #include <wx/grid.h>
 #include <wx/sizer.h>
 #include <wx/string.h>
