+++ src/import/Import.cpp
@@ -35,6 +35,10 @@ and ImportLOF.cpp.
 
 
 
+#include "../Audacity.h"
+#include "Import.h"
+#include "ImportPlugin.h"
+
 #include <wx/textctrl.h>
 #include <wx/msgdlg.h>
 #include <wx/string.h>
@@ -44,11 +48,8 @@ and ImportLOF.cpp.
 #include <wx/arrimpl.cpp>
 #include <wx/listimpl.cpp>
 #include "../ShuttleGui.h"
-#include "../Audacity.h"
 #include "../Project.h"
 
-#include "Import.h"
-#include "ImportPlugin.h"
 #include "ImportPCM.h"
 #include "ImportMP3.h"
 #include "ImportOGG.h"
