+++ src/widgets/ProgressDialog.cpp
@@ -28,6 +28,7 @@
 
 #include "../Audacity.h"
 
+#include "../MemoryX.h"
 #include <wx/defs.h>
 #include <wx/app.h>
 #include <wx/button.h>
